+++ e2e-tests/tests/add-entity-page.spec.ts
@@ -11,11 +11,15 @@ studioTest("can add an entity page", async ({ page, studioPage }) => {
   const pageInTree = page.getByText("EntityPage");
   await expect(pageInTree).toHaveCount(0);
 
+  const addPageSection = studioPage.addPageSection;
+  await addPageSection.selectPageType(true);
+  await addPageSection.enterBasicPageData("EntityPage", "entity-page");
+  await addPageSection.enterStreamScope({}, true);
+  await addPageSection.selectLayout();
+
   // TODO: Specify a stream scope once we can supply an API key for populating
   // the store with account content
-  await studioPage.addEntityPage("EntityPage", {}, "entity-page");
   await expect(pageInTree).toHaveCount(1);
-  await studioPage.takePageScreenshotAfterImgRender();
   await studioPage.saveButton.click();
   const expectedPagePath = studioPage.getPagePath("EntityPage");
   await expect(expectedPagePath).toHaveContents(expectedPage);
